% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{adjust_and_split}
\alias{adjust_and_split}
\title{Adjust list of p-value matrices}
\usage{
adjust_and_split(p_val_list, adjust_method="BY")
}
\arguments{
\item{p_val_list}{list of matrices containing uncorrected p-vales}

\item{adjust_method}{method for p-value correction. Default: "BY". For details see \link{p.adjust}}
}
\value{
a \code{list}
}
\description{
Adjusts all p-values in the list and separates them finally into the same matrix structure
}
\author{
Alexander Gabel
}
