% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_graft_formation}
\alias{barplot_graft_formation}
\title{Barplot of up- and down- regulated differentially expressed genes}
\usage{
barplot_graft_formation(count_mat, pval_mat, ylab="# genes intersecting", pval_threshold = 0.05, main)
}
\arguments{
\item{count_mat}{number of genes during each time point for each condition}

\item{pval_mat}{matrix of p-values for each time point and each treatment. Indicating if ratio of up- and down- regulated genes is significant compared to the background.}

\item{ylab}{defines the label of the y-axis}

\item{pval_threshold}{p-value threshold to indocate if a p-value is significant}

\item{main}{defines the title of the plot}
}
\value{
a barplot
}
\description{
Plots the relative or absolute number of up- and down- regulated genes and a asterisks if the ratio is significant
}
\author{
Alexander Gabel
}
