% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barplot_up_down}
\alias{barplot_up_down}
\title{Barplot of up- and down- regulated differentially expressed genes}
\usage{
barplot_up_down(up_mat, down_mat, p_val_mat, names.arg, main, labels)
}
\arguments{
\item{up_mat, down_mat}{(relative) number of up or down regulated genes}

\item{p_val_mat}{matrix of p-values for each time point and each treatment. Indicating if ratio of up- and down- regulated genes is significant compared to the background.}

\item{ylim}{defines the limits of the y-axis}

\item{names.arg}{defines the names of the bars. Details: \link{barplot}}

\item{main}{defines the title of the plot}

\item{labels}{defines the labels of the x-axis}

\item{ylab}{defines the label of the y-axis}
}
\value{
a barplot
}
\description{
Plots the relative or absolute number of up- and down- regulated genes and a asterisks if the ratio is significant
}
\author{
Alexander Gabel
}
