% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{Compute_GO_Enrichment}
\alias{Compute_GO_Enrichment}
\title{Plot dendrogram colored by sample definition}
\usage{
Compute_GO_Enrichment(geneUniverse, selectedGeneIds, pvalueCutoff=0.05, ontology="BP")
}
\arguments{
\item{geneUniverse}{\code{vector} containing all gene ids of the whole dataset}

\item{selectedGeneIds}{\code{vector} containing subset of gene ids to check if their GO terms are enriched}

\item{pvalueCutoff}{numeric value defining the p-value cutoff for the adjusted p-values to call a GO term significantly enriched (default: 0.05)}

\item{ontology}{character string defining the GO ontology and must be one of "BP", "CC", or "MF"}
}
\value{
a \code{list}
}
\description{
Does a hierarchical clustering of the data and draws a dendrogam with colored branches and leaved based on the sample definition.
}
\author{
Alexander Gabel
}
