% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{do_fisher_up_down}
\alias{do_fisher_up_down}
\title{Fisher test for transcriptional overlap}
\usage{
do_fisher_up_down(fc.list, fg.ids, ml.list.up, ml.list.down, fc.threshold, ml.threshold=0.9, alternative = "two.sided")
}
\arguments{
\item{fc.list}{list containing the log2-foldchange matrices of treatment [grafted|separated] [top|bottom] vs. intact}

\item{fg.ids}{vector containing gene ids which shall be used for overlap analysis}

\item{ml.list.up}{list containing matrices with marginal likelihoods if a gene is up regulated}

\item{ml.list.down}{list containing matrices with marginal likelihoods if a gene is down regulated}

\item{fc.threshold}{log2 foldchange threshold to define if a gene is differentially expressed}

\item{ml.threshold}{marginal likelihood threshold to define if a gene is differentially expressed}

\item{alternative}{character string indicating the alternative hypothesis and must be one of "two.sided" (default), "greater", or "less". Details \link{fisher.test}}
}
\value{
a \code{list}
}
\description{
Calculate significance if ratio of differentially expressed genes from a given gene set is significantly different to the ratio of differentially expressed genes in the complete dataset
}
\author{
Alexander Gabel
}
