% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp.data}
\alias{exp.data}
\title{Expression data set of grafting.}
\format{A data frame with 27416 rows and 82 variables:
\describe{
  \item{0A Intact}{length normed counts}
  \item{0B Intact}{length normed counts}
  ...
  \item{240B Grafted top}{length normed counts}
}}
\usage{
exp.data
}
\description{
A dataset containing the length normed count data from intact, grafted
and separated samples.
}
\keyword{datasets}
