% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fc.list}
\alias{fc.list}
\title{List of foldchanges}
\format{A list with 4 elements
\describe{
  \item{fc.list$`Grafted bottom`}{matrix of foldchanges}
  \item{fc.list$`Grafted top`}{matrix of foldchanges}
  \item{fc.list$`Seperated bottom`}{matrix of foldchanges}
  \item{fc.list$`Seperated top`}{matrix of foldchanges}
}}
\usage{
fc.list
}
\description{
A dataset containing a list of foldchanges for each time point
of grafted and separarted samples compared to intact samples
}
\keyword{datasets}
