% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groupS}
\alias{groupS}
\title{Vector containing groups for summarized samples}
\format{A numeric vector indicating the relation to a certain group}
\usage{
groupS
}
\description{
Vector containing groups for summarized samples
}
\keyword{datasets}
