% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{intersect.analysis}
\alias{intersect.analysis}
\title{Intersection analysis}
\usage{
intersect.analysis(gene.list, geneUniverse, alternative="greater")
}
\arguments{
\item{gene.list}{list containing vectors of gene ids}

\item{geneUniverse}{all gene ids in the dataset}

\item{alternative}{character string indicating the alternative hypothesis and must be one of "two.sided", "greater" (default), or "less". Details \link{fisher.test}}
}
\value{
a \code{plot}
}
\description{
Calculate if overlap between different gene list ids is significant
}
\author{
Alexander Gabel
}
