% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ml.list.down}
\alias{ml.list.down}
\title{List of marginal likelihoods of downregulated genes compared to intact}
\format{A list with 4 elements
\describe{
  \item{ml.list.down$`Grafted bottom`}{matrix of foldchanges}
  \item{ml.list.down$`Grafted top`}{matrix of foldchanges}
  \item{ml.list.down$`Seperated bottom`}{matrix of foldchanges}
  \item{ml.list.down$`Seperated top`}{matrix of foldchanges}
}}
\usage{
ml.list.down
}
\description{
A dataset containing a list of marginal likelihoods for each time point
of grafted and separarted samples compared to intact samples
}
\keyword{datasets}
