% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_colored_dendrogram}
\alias{plot_colored_dendrogram}
\title{Plot dendrogram colored by sample definition}
\usage{
plot_colored_dendrogram(exp_data,groups, method="pearson", link="average", do.legend=F, unClusteredBranchColor="black", legend.pos = "topright", ...)
}
\arguments{
\item{exp_data}{expression matrix, columns represent the samples and rows the genes (or transcripts)}

\item{groups}{numeric vector describing the relationships of columns to certain groups (like biological replicates)}

\item{method}{character string defining which correlation coefficient should be used as similarity measure pearson (default), spearman, or kendall}

\item{link}{character sting defining which agglomeration method to be used. Default: "average". For more information see parameter method at \link{hclust}}

\item{do.legend}{boolean value defining if a color legend should be ploted}

\item{unClusteredBranchColor}{character string defining the color to use if two brnaches belong to different sample types}

\item{legend.pos}{character string defining the position of the legend e.g. "topright" (default), "topleft",... \link{legend}}

\item{...}{optional plot parameters \link{plot}}
}
\value{
a \code{plot}
}
\description{
Does a hierarchical clustering of the data and draws a dendrogam with colored branches and leaved based on the sample definition.
}
\author{
Alexander Gabel
}
