% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyses.R
\name{test_sym_asym}
\alias{test_sym_asym}
\title{Fisher test to check significance of overlap}
\usage{
test_sym_asym(gene_set_1, gene_set_2, number_genes_complete, alternative = "greater")
}
\arguments{
\item{gene_set_1}{character vector containing gene ids of first gene set}

\item{gene_set_2}{character vector containing gene ids of second gene set}

\item{number_genes_complete}{number of genes in the complete dataset}

\item{alternative}{character string indicating the alternative hypothesis and must be one of "two.sided" (default), "greater", or "less". Details \link{fisher.test}}
}
\description{
Calculate significance of overlap between two gene sets
}
\author{
Alexander Gabel
}
